+++ src/Enchantments/ReturnToHandEnchantment.cs
@@ -14,15 +14,14 @@ public sealed class ReturnToHandEnchantment : ModEnchantmentTemplate, IRewardEnc
 
 	public override bool CanEnchantCardType(CardType cardType) => cardType != CardType.Power;
 
-	public override (PileType, CardPilePosition) ModifyCardPlayResultPileTypeAndPosition(
+	public override CardLocation ModifyCardPlayResultLocation(
 		CardModel card,
 		bool isAutoPlay,
 		ResourceInfo resources,
-		PileType pileType,
-		CardPilePosition position)
+		CardLocation cardLocation)
 	{
-		if (Card != card || pileType != PileType.Discard)
-			return (pileType, position);
-		return (PileType.Hand, position);
+		if (Card != card || cardLocation.pileType != PileType.Discard)
+			return cardLocation;
+		return new CardLocation(cardLocation.player, PileType.Hand, cardLocation.position);
 	}
 }